Initial commit.
moot utility works on it's own for now, but has a problem with a bug in tar
that should likely be fixed with a speical invocation of libuid, hence the
generic name. libuid is planned to be mostly used for groot, so that it can
fake out getuid and geteuid calls on a program specific basis. Planning on
passing env variables as arguments to the library.